msm/sde/rotator: Remove unneeded PM QoS requests
When the rotator is actually used (still an unsolved question in
computer science), these PM QoS requests block some CPUs in the LITTLE
cluster from entering deep idle because the driver assumes that display
rotating work occurs on a hardcoded set of CPUs, which is false. We
already have the IRQ PM QoS machinery for display rendering operations
that actually matter, so this cruft is unneeded.

Change-Id: I30542fa3009dd46ff0c21058f14c8923d7fcb892
Signed-off-by: Sultan Alsawaf <sultan@kerneltoast.com>
Signed-off-by: firebird11 <hbgassel@gmail.com>